Project with Hibernate (#4)
* Refactor to hibernate: admin(delete, wa mutual delete), user (profile update)

* deleted comments, edit Article form

* added sm Tests

* bugfix4 (after review): controllers fix

* added generic class

* Add article added

* bugfix (unusued clean)

+++ service-module/src/main/java/by/tut/mdcatalog/project2/service/ReviewService.java
@@ -8,5 +8,5 @@ public interface ReviewService {
 
     List<ReviewDTO> getReviews();
 
-    void deleteReviews(int[] ids);
+    void deleteReviews(Long[] ids);
 }
